代码调整

+++ scms/src/main/java/com/ashuo/scms/controller/SyslogController.java
@@ -15,6 +15,7 @@ import org.apache.commons.lang.StringUtils;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.apache.shiro.authz.annotation.RequiresRoles;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.*;
 
 /**
@@ -31,6 +32,8 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/syslog")
 public class SyslogController {
 
+    @Value("${meet.you.year}")
+    private String meetYouYear;
     public  static  boolean systemStatus=true;
 
     @Autowired
@@ -73,26 +76,29 @@ public class SyslogController {
         return ServerResponse.createBySuccess(systemStatus);
     }
 
-    @ApiOperation("清空项目排名分数运动员数据")
-    @DeleteMapping("/d")
-    @RequiresRoles(value = {"1"})
-    public ServerResponse resetRangkingScoreAthlete(Integer itemId) {
-        int effNum = 0;
-        try {
-            effNum = syslogService.removeAllRankingScoreAthlete();
-        } catch (Exception e) {
-            return ServerResponse.createByErrorCodeMessage(400, "清空失败");
-        }
-        if (effNum == 0) {
-            return ServerResponse.createByErrorCodeMessage(400, "清空失败");
-        }
-        return ServerResponse.createBySuccessMessage("清空成功");
-    }
+//    @ApiOperation("清空项目排名分数运动员数据")
+//    @DeleteMapping("/d")
+//    @RequiresRoles(value = {"1"})
+//    public ServerResponse resetRangkingScoreAthlete(Integer itemId) {
+//        int effNum = 0;
+//        try {
+//            effNum = syslogService.removeAllRankingScoreAthlete();
+//        } catch (Exception e) {
+//            return ServerResponse.createByErrorCodeMessage(400, "清空失败");
+//        }
+//        if (effNum == 0) {
+//            return ServerResponse.createByErrorCodeMessage(400, "清空失败");
+//        }
+//        return ServerResponse.createBySuccessMessage("清空成功");
+//    }
 
     @ApiOperation("清空系统数据")
     @DeleteMapping("/resetAllData")
     @RequiresRoles(value = {"1"})
-    public ServerResponse resetAllData(Integer itemId) {
+    public ServerResponse resetAllData(String comfirmPassword) {
+        if (!comfirmPassword.equals(meetYouYear)){
+            return ServerResponse.createByErrorCodeMessage(400, "密码输入错误");
+        }
         int effNum = 0;
         try {
             effNum = syslogService.removeAllData();
